[Add] スクロールの上下でトップページボタンのクラスを切り替える機能を実装
