+++ app/javascript/packs/application.js
@@ -15,4 +15,4 @@ Rails.start()
 ActiveStorage.start()
 
 import '../features/hamburger-toggle'
-import '../features/header-scroll-trigger'
+import '../features/class-toggle-scroller'
